requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.4.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.1
triton==2.1.0
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
